+++ ft_printf.h
@@ -12,6 +12,6 @@ void	ft_putnbr(int i, int *counter);
 void	ft_putnbr_unsigned(unsigned i, char *base, int *counter);
 void	ft_hexa(unsigned i, char *base, int *counter);
 int		ft_printf(const char *str, ...);
-void	ft_pointer_adress(unsigned i, char *base, int *counter);
+void	ft_pointer_adress(unsigned long i, char *base, int *counter);
 
 #endif
